♻️ Trekk valideringsfunksjoner for vedtaksperioder tilsyn barn ut i Utils fil for å enklere kunne teste
